requirements.txt
appdirs==1.4.4
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
cssselect==1.2.0
Django==4.1.4
djangorestframework==3.14.0
fake-useragent==1.1.1
idna==3.4
importlib-metadata==5.1.0
lxml==4.9.1
parse==1.19.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
pytz==2022.6
requests==2.28.1
requests-html==0.10.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.7
urllib3==1.26.13
w3lib==2.1.1
websockets==10.4
zipp==3.11.0
